remove dmsgget command | add example script
